Offload CrossPlatLock responsiblity to portalocker.
Turns out that opening files with an exclusive read/write lock is more
complicated than it ought to be, and there isn't a portable way of doing
it using the standard library.

Luckily, there is a library that handles this. From brief analysis, it
seems like a stable enough library to take a dependency on.